--- client/src/components/common/addModal.tsx
@@ -1,1 +1,0 @@
-export {};
